+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/file_processor_rvt.h
@@ -39,7 +39,7 @@ namespace repo {
 					FileProcessorRvt(const std::string& inputFile, GeometryCollector* geoCollector) : FileProcessor(inputFile, geoCollector) {};
 					~FileProcessorRvt() override;
 
-					int readFile() override;
+					uint8_t readFile() override;
 
 				private:
 					//.. adjust triangulation options
